Fixed JSON Bug + Data Retrival.
Had a bug where when I submitted a form, it got pushed to both json files for both employees and projects.

Forms for both add employees and projects are created and do a  post request to the server side.

Implemented some basic CSS which I hate.
Responsive design is still yet to come.