Added A Fixed Unique Page, along with A Timer Which needs to be configured to JSON
Okay okay so for notes I added in a unique page and I wanted to put some thoughts.

With every login, with a supposed correct username and password (they can make this up for now -- later I will add in a login -- same json file??) where when the user wants to "retire" or even like change their status of if theyre on leave etc. then they can do so.

Now also, I want to fix up the tasks section, it seems with every new task it is being overwritten , and im honestly tired so Im dumb when im tired. Im thinking of creating a new key, with an object. with every key will be the id number of the project which isunique, and the object will contain the specific keys and values of the add a task form.

Same idea with the users, rahter than it being a super larga rray with objects, thinking of doing a key with the eprsons name and in the object have their name, username, generated pin code, etc!